refactor: consolidate backend rendering and segregate frontend presenter hooks
