+++ app/web/feeds/renderer.rb
@@ -26,7 +26,83 @@ module Html2rss
           - Reach out to the site owner if access is restricted
         DESC
 
+        JSON_FEED = :json_feed
+        RSS = :rss
+
+        JSON_CONTENT_TYPE = 'application/feed+json'
+        RSS_CONTENT_TYPE = 'application/xml'
+
+        PATH_FORMATS = {
+          '.json' => JSON_FEED,
+          '.rss' => RSS,
+          '.xml' => RSS
+        }.freeze
+
+        JSON_MEDIA_TYPES = [
+          'application/feed+json',
+          'application/json'
+        ].freeze
+
+        RSS_MEDIA_TYPES = [
+          'application/rss+xml',
+          'application/xml',
+          'text/xml'
+        ].freeze
+
+        MediaRange = Data.define(:type, :subtype, :quality, :position) do
+          # @return [Integer]
+          def specificity
+            return 0 if type == '*' && subtype == '*'
+            return 1 if subtype == '*'
+
+            2
+          end
+
+          # @param candidate [String]
+          # @return [Boolean]
+          def matches?(candidate)
+            candidate_type, candidate_subtype = candidate.downcase.split('/', 2)
+            return true if type == '*' && subtype == '*'
+            return candidate_type == type if subtype == '*'
+
+            candidate_type == type && candidate_subtype == subtype
+          end
+        end
+
         class << self
+          # Renders a RenderResult and configures the HTTP response headers and status.
+          #
+          # @param result [Html2rss::Web::Feeds::Contracts::RenderResult]
+          # @param response [Rack::Response]
+          # @param request [Rack::Request]
+          # @return [String] serialized feed representation
+          def render(result, response:, request:)
+            format = for_request(request)
+
+            response.status = status_for(result.status)
+            response['Content-Type'] = content_type(format)
+            apply_cache_headers(response, result)
+            ::Html2rss::Web::HttpCache.vary(response, 'Accept')
+
+            call(result, format: format)
+          end
+
+          # Renders an error feed and configures the HTTP response headers and status.
+          #
+          # @param message [String] the error message to display in the feed.
+          # @param response [Rack::Response] the Rack response.
+          # @param request [Rack::Request] the Rack request.
+          # @param format [Symbol, nil] explicit format (:rss or :json_feed), or nil to negotiate.
+          # @return [String] serialized error feed representation
+          def render_error(message, response:, request:, format: nil)
+            resolved_format = format || for_request(request)
+
+            response['Content-Type'] = content_type(resolved_format)
+            ::Html2rss::Web::HttpCache.expires_now(response)
+
+            call_error(message: message, format: resolved_format)
+          end
+
           # Renders a RenderResult into the requested format.
           #
           # @param result [Html2rss::Web::Feeds::Contracts::RenderResult]
@@ -53,7 +129,7 @@ module Html2rss
             desc = "Failed to generate feed: #{message}"
             timestamp = Time.now.utc
 
-            if format == FeedResponseFormat::JSON_FEED
+            if format == JSON_FEED
               JSON.generate({
                               version: JSON_FEED_VERSION,
                               title: title,
@@ -79,13 +155,154 @@ module Html2rss
             end
           end
 
+          # @param request [Rack::Request]
+          # @return [Symbol] negotiated feed format.
+          def for_request(request)
+            from_path(request_path(request)) || from_accept(accept_header(request)) || RSS
+          end
+
+          # @param path [String]
+          # @return [Symbol, nil] format implied by known extension.
+          def from_path(path)
+            PATH_FORMATS.each do |suffix, format|
+              return format if path.end_with?(suffix)
+            end
+
+            nil
+          end
+
+          # @param value [String]
+          # @return [String] input without a known feed extension suffix.
+          def strip_known_extension(value)
+            string = value.to_s
+
+            PATH_FORMATS.each_key do |suffix|
+              return string.delete_suffix(suffix) if string.end_with?(suffix)
+            end
+
+            string
+          end
+
+          # @param format [Symbol]
+          # @return [String] HTTP content type for the negotiated format.
+          def content_type(format)
+            format == JSON_FEED ? JSON_CONTENT_TYPE : RSS_CONTENT_TYPE
+          end
+
+          # Parses Accept header and returns the preferred format based on priority.
+          #
+          # @param accept_header [String, nil]
+          # @return [Symbol, nil] preferred format (:json_feed, or nil meaning fallback to rss)
+          def from_accept(accept_header)
+            media_ranges = parse_accept(accept_header)
+            return nil if media_ranges.empty?
+
+            json_score = best_score(media_ranges, JSON_MEDIA_TYPES)
+            rss_score = best_score(media_ranges, RSS_MEDIA_TYPES)
+
+            return nil unless json_score
+            return JSON_FEED if rss_score.nil?
+
+            (json_score <=> rss_score)&.positive? ? JSON_FEED : nil
+          end
+
           private
+
+          # @param response [Rack::Response]
+          # @param result [Html2rss::Web::Feeds::Contracts::RenderResult]
+          # @return [void]
+          def apply_cache_headers(response, result)
+            return ::Html2rss::Web::HttpCache.expires_now(response) if result.status == :error
+
+            ::Html2rss::Web::HttpCache.expires(response, result.ttl_seconds, cache_control: 'public')
+          end
+
+          # @param status [Symbol]
+          # @return [Integer]
+          def status_for(status)
+            return 200 if status == :ok
+            return 422 if status == :empty
+
+            500
+          end
+
+          # @param request [Rack::Request]
+          # @return [String]
+          def request_path(request)
+            path = request.respond_to?(:env) ? request.env['PATH_INFO'] : nil
+            return path.to_s unless request_path_fallback?(request, path)
+
+            request.path_info.to_s
+          end
+
+          # @param request [Rack::Request]
+          # @return [String, nil]
+          def accept_header(request)
+            return request.get_header('HTTP_ACCEPT') unless request.respond_to?(:env)
+
+            request.env['HTTP_ACCEPT'] || request.get_header('HTTP_ACCEPT')
+          end
+
+          # @param request [Rack::Request]
+          # @param path [String, nil]
+          # @return [Boolean]
+          def request_path_fallback?(request, path)
+            path.to_s.empty? && request.respond_to?(:path_info)
+          end
+
+          # @param accept_header [String, nil]
+          # @return [Array<MediaRange>]
+          def parse_accept(accept_header)
+            accept_header.to_s.split(',').filter_map.with_index do |raw_range, position|
+              build_media_range(raw_range, position)
+            end
+          end
+
+          # @param raw_range [String]
+          # @param position [Integer]
+          # @return [MediaRange, nil]
+          def build_media_range(raw_range, position)
+            media_type, *parameter_parts = raw_range.strip.downcase.split(';')
+            type, subtype = media_type.to_s.split('/', 2)
+            return if type.to_s.empty? || subtype.to_s.empty?
+
+            MediaRange.new(
+              type: type,
+              subtype: subtype,
+              quality: extract_quality(parameter_parts),
+              position: position
+            )
+          end
+
+          # @param parameter_parts [Array<String>]
+          # @return [Float]
+          def extract_quality(parameter_parts)
+            raw_value = parameter_parts
+                        .map(&:strip)
+                        .find { |part| part.start_with?('q=') }
+                        &.split('=', 2)
+                        &.last
+            quality = raw_value ? Float(raw_value) : 1.0
+            quality.clamp(0.0, 1.0)
+          rescue ArgumentError
+            1.0
+          end
+
+          # @param media_ranges [Array<MediaRange>]
+          # @param candidates [Array<String>]
+          # @return [Array(Float, Integer, Integer), nil]
+          def best_score(media_ranges, candidates)
+            media_ranges
+              .filter { |range| range.quality.positive? && candidates.any? { |candidate| range.matches?(candidate) } }
+              .map { |range| [range.quality, range.specificity, -range.position] }
+              .max
+          end
 
           # @param result [Html2rss::Web::Feeds::Contracts::RenderResult]
           # @param format [Symbol]
           # @return [String]
           def render_success(result, format)
-            if format == FeedResponseFormat::JSON_FEED
+            if format == JSON_FEED
               serialize_json_feed(result.payload.feed)
             else
               result.payload.feed.to_s
@@ -104,7 +321,7 @@ module Html2rss
             desc = empty_feed_description(url, strategy)
             timestamp = Time.now.utc
 
-            if format == FeedResponseFormat::JSON_FEED
+            if format == JSON_FEED
               JSON.generate({
                 version: JSON_FEED_VERSION,
                 title: title,
@@ -178,7 +395,7 @@ module Html2rss
           # @param timestamp [Time]
           # @return [String]
           def build_rss(title:, description:, link: nil, items: [], timestamp: nil)
-            RSS::Maker.make('2.0') do |maker|
+            ::RSS::Maker.make('2.0') do |maker|
               # Apply stylesheets
               stylesheets = Html2rss.configuration.stylesheets.map do |s|
                 Html2rss::RssBuilder::Stylesheet.new(**s)
